+++ CHANGELOG.md
@@ -12,3 +12,9 @@ Every sprint's progress reflected here.
 - Registration Page [#29](https://github.com/mulla028/T9-Final_Project/pull/29)
 - Forgot Password logic and UI/UX using `nodemailer` Added [#28](https://github.com/mulla028/T9-Final_Project/pull/28)
 - Login Page logic and UI implemented using OAuth [#23](https://github.com/mulla028/T9-Final_Project/pull/23)
+
+## Sprint 2
+
+### Added
+
+- Added the search feature with API in homepage and map page[#38](https://github.com/mulla028/T9-Final_Project/issues/38)